// src/app/models/database.types.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "14.5"
  }
  public: {
    Tables: {
// ...
      program_exercises: {
        Row: {
          default_reps: string | null
          default_sets: number | null
          exercise_id: number
          id: number
          notes: string | null
          position: number
          program_id: number
        }
        Insert: {
          default_reps?: string | null
          default_sets?: number | null
          exercise_id: number
          id?: number
          notes?: string | null
          position?: number
          program_id: number
        }
        Update: {
          default_reps?: string | null
          default_sets?: number | null
          exercise_id?: number
          id?: number
          notes?: string | null
          position?: number
          program_id?: number
        }
        Relationships: [
          {
            foreignKeyName: "program_exercises_exercise_id_fkey"
            columns: ["exercise_id"]
            isOneToOne: false
            referencedRelation: "exercises"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "program_exercises_program_id_fkey"
            columns: ["program_id"]
            isOneToOne: false
            referencedRelation: "programs"
            referencedColumns: ["id"]
          },
        ]
      }
      programs: {
        Row: {
          description: string | null
          id: number
          name: string
        }
        Insert: {
          description?: string | null
          id?: number
          name: string
        }
        Update: {
          description?: string | null
          id?: number
          name?: string
        }
        Relationships: []
      }
// ...
      workouts: {
        Row: {
          id: number
          notes: string | null
          performed_at: string
          program_id: number | null
        }
        Insert: {
          id?: number
          notes?: string | null
          performed_at?: string
          program_id?: number | null
        }
        Update: {
          id?: number
          notes?: string | null
          performed_at?: string
          program_id?: number | null
        }
        Relationships: [
          {
            foreignKeyName: "workouts_program_id_fkey"
            columns: ["program_id"]
            isOneToOne: false
            referencedRelation: "programs"
            referencedColumns: ["id"]
          },
        ]
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}